// angular/dynamic-ui/test/primeng/src/dynamic-ui-primeng-form.const.d.ts
export declare const enum PrimeNGFormControlType {
    Array = 1,
    AutoComplete = 2,
    Calendar = 3,
    Checkbox = 4,
    Chips = 5,
    Dropdown = 6,
    Editor = 7,
    Group = 8,
    Input = 9,
    InputSwitch = 10,
    MultiSelect = 11,
    RadioGroup = 12,
    Rating = 13,
    Slider = 14,
    TextArea = 15,
    Columns = 16,
}
export declare const PRIME_NG_VIEW_CHILD_SELECTOR: string;
export declare const PRIME_NG_AUTOCOMPLETE_TEMPLATE_DIRECTIVES: {
    itemTemplate: string;
    selectedItemTemplate: string;
};
export declare const PRIME_NG_CHIPS_TEMPLATE_DIRECTIVES: {
    itemTemplate: string;
};
export declare const PRIME_NG_DROPDOWN_LIST_TEMPLATE_DIRECTIVES: {
    itemTemplate: string;
};